refactor cli error boundaries and workspace commands
