+++ src/commands/repository.ts
@@ -8,7 +8,6 @@ import type {
   UpdateRepo,
 } from "../api/types.ts";
 import { applyFilters } from "../utils/filter.ts";
-import { handleCliError } from "../utils/error-handler.ts";
 import { getRepositoryId } from "../utils/repository-resolver.ts";
 
 export const repositoryCommand = new Command()
@@ -25,48 +24,41 @@ repositoryCommand
   .option("--path <path:string>", "Filter by repository path")
   .option("--json", "Output as JSON")
   .action(async (options) => {
-    try {
-      const client = await ApiClient.create();
-      let repos = await client.listRepos();
+    const client = await ApiClient.create();
+    let repos = await client.listRepos();
 
-      // Build filter object from provided options
-      const filters: Record<string, unknown> = {};
-      if (options.name !== undefined) {
-        filters.name = options.name;
-      }
-      if (options.path !== undefined) {
-        filters.path = options.path;
-      }
-
-      // Apply filters
-      repos = applyFilters(repos, filters);
-
-      if (options.json) {
-        console.log(JSON.stringify(repos, null, 2));
-        return;
-      }
-
-      if (repos.length === 0) {
-        console.log("No repositories found.");
-        return;
-      }
-
-      const table = new Table()
-        .header(["ID", "Name", "Display Name", "Path"])
-        .body(
-          repos.map((r) => [
-            r.id,
-            r.name,
-            r.display_name || "-",
-            r.path,
-          ]),
-        );
-
-      table.render();
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    const filters: Record<string, unknown> = {};
+    if (options.name !== undefined) {
+      filters.name = options.name;
     }
+    if (options.path !== undefined) {
+      filters.path = options.path;
+    }
+
+    repos = applyFilters(repos, filters);
+
+    if (options.json) {
+      console.log(JSON.stringify(repos, null, 2));
+      return;
+    }
+
+    if (repos.length === 0) {
+      console.log("No repositories found.");
+      return;
+    }
+
+    const table = new Table()
+      .header(["ID", "Name", "Display Name", "Path"])
+      .body(
+        repos.map((r) => [
+          r.id,
+          r.name,
+          r.display_name || "-",
+          r.path,
+        ]),
+      );
+
+    table.render();
   });
 
 // Show repository
@@ -76,50 +68,45 @@ repositoryCommand
   .arguments("[id:string]")
   .option("--json", "Output as JSON")
   .action(async (options, id?: string) => {
-    try {
-      const client = await ApiClient.create();
-      const repoId = await getRepositoryId(id, client);
-      const repo = await client.getRepo(repoId);
+    const client = await ApiClient.create();
+    const repoId = await getRepositoryId(id, client);
+    const repo = await client.getRepo(repoId);
 
-      if (options.json) {
-        console.log(JSON.stringify(repo, null, 2));
-        return;
-      }
-
-      console.log(`ID:                   ${repo.id}`);
-      console.log(`Name:                 ${repo.name}`);
-      console.log(`Display Name:         ${repo.display_name || "-"}`);
-      console.log(`Path:                 ${repo.path}`);
-      if (repo.setup_script) {
-        console.log(`Setup Script:         ${repo.setup_script}`);
-      }
-      if (repo.cleanup_script) {
-        console.log(`Cleanup Script:       ${repo.cleanup_script}`);
-      }
-      if (repo.archive_script) {
-        console.log(`Archive Script:       ${repo.archive_script}`);
-      }
-      if (repo.copy_files) {
-        console.log(`Copy Files:           ${repo.copy_files}`);
-      }
-      console.log(
-        `Parallel Setup:       ${repo.parallel_setup_script ? "Yes" : "No"}`,
-      );
-      if (repo.dev_server_script) {
-        console.log(`Dev Server Script:    ${repo.dev_server_script}`);
-      }
-      if (repo.default_target_branch) {
-        console.log(`Default Target:       ${repo.default_target_branch}`);
-      }
-      if (repo.default_working_dir) {
-        console.log(`Default Working Dir:  ${repo.default_working_dir}`);
-      }
-      console.log(`Created:              ${repo.created_at}`);
-      console.log(`Updated:              ${repo.updated_at}`);
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    if (options.json) {
+      console.log(JSON.stringify(repo, null, 2));
+      return;
     }
+
+    console.log(`ID:                   ${repo.id}`);
+    console.log(`Name:                 ${repo.name}`);
+    console.log(`Display Name:         ${repo.display_name || "-"}`);
+    console.log(`Path:                 ${repo.path}`);
+    if (repo.setup_script) {
+      console.log(`Setup Script:         ${repo.setup_script}`);
+    }
+    if (repo.cleanup_script) {
+      console.log(`Cleanup Script:       ${repo.cleanup_script}`);
+    }
+    if (repo.archive_script) {
+      console.log(`Archive Script:       ${repo.archive_script}`);
+    }
+    if (repo.copy_files) {
+      console.log(`Copy Files:           ${repo.copy_files}`);
+    }
+    console.log(
+      `Parallel Setup:       ${repo.parallel_setup_script ? "Yes" : "No"}`,
+    );
+    if (repo.dev_server_script) {
+      console.log(`Dev Server Script:    ${repo.dev_server_script}`);
+    }
+    if (repo.default_target_branch) {
+      console.log(`Default Target:       ${repo.default_target_branch}`);
+    }
+    if (repo.default_working_dir) {
+      console.log(`Default Working Dir:  ${repo.default_working_dir}`);
+    }
+    console.log(`Created:              ${repo.created_at}`);
+    console.log(`Updated:              ${repo.updated_at}`);
   });
 
 // Register existing repository
@@ -129,36 +116,31 @@ repositoryCommand
   .option("--path <path:string>", "Path to the git repository")
   .option("--display-name <name:string>", "Display name for the repository")
   .action(async (options) => {
-    try {
-      let path = options.path;
-      let displayName = options.displayName;
+    let path = options.path;
+    let displayName = options.displayName;
 
-      if (!path) {
-        path = await Input.prompt("Repository path:");
-      }
-
-      if (displayName === undefined) {
-        displayName = await Input.prompt({
-          message: "Display name (optional):",
-          default: "",
-        });
-      }
-
-      const request: RegisterRepoRequest = {
-        path,
-        display_name: displayName || null,
-      };
-
-      const client = await ApiClient.create();
-      const repo = await client.registerRepo(request);
-
-      console.log(`Repository registered successfully!`);
-      console.log(`ID: ${repo.id}`);
-      console.log(`Name: ${repo.name}`);
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    if (!path) {
+      path = await Input.prompt("Repository path:");
     }
+
+    if (displayName === undefined) {
+      displayName = await Input.prompt({
+        message: "Display name (optional):",
+        default: "",
+      });
+    }
+
+    const request: RegisterRepoRequest = {
+      path,
+      display_name: displayName || null,
+    };
+
+    const client = await ApiClient.create();
+    const repo = await client.registerRepo(request);
+
+    console.log(`Repository registered successfully!`);
+    console.log(`ID: ${repo.id}`);
+    console.log(`Name: ${repo.name}`);
   });
 
 // Initialize new repository
@@ -168,33 +150,28 @@ repositoryCommand
   .option("--parent-path <path:string>", "Parent directory path")
   .option("--folder-name <name:string>", "Folder name for the new repository")
   .action(async (options) => {
-    try {
-      let parentPath = options.parentPath;
-      let folderName = options.folderName;
+    let parentPath = options.parentPath;
+    let folderName = options.folderName;
 
-      if (!parentPath) {
-        parentPath = await Input.prompt("Parent directory path:");
-      }
-
-      if (!folderName) {
-        folderName = await Input.prompt("Folder name:");
-      }
-
-      const request: InitRepoRequest = {
-        parent_path: parentPath,
-        folder_name: folderName,
-      };
-
-      const client = await ApiClient.create();
-      const repo = await client.initRepo(request);
-
-      console.log(`Repository initialized successfully!`);
-      console.log(`ID: ${repo.id}`);
-      console.log(`Path: ${repo.path}`);
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    if (!parentPath) {
+      parentPath = await Input.prompt("Parent directory path:");
     }
+
+    if (!folderName) {
+      folderName = await Input.prompt("Folder name:");
+    }
+
+    const request: InitRepoRequest = {
+      parent_path: parentPath,
+      folder_name: folderName,
+    };
+
+    const client = await ApiClient.create();
+    const repo = await client.initRepo(request);
+
+    console.log(`Repository initialized successfully!`);
+    console.log(`ID: ${repo.id}`);
+    console.log(`Path: ${repo.path}`);
   });
 
 // Update repository
@@ -219,50 +196,45 @@ repositoryCommand
     "Default working directory for workspace execution",
   )
   .action(async (options, id?: string) => {
-    try {
-      const update: UpdateRepo = {};
+    const update: UpdateRepo = {};
 
-      if (options.displayName !== undefined) {
-        update.display_name = options.displayName || null;
-      }
-      if (options.setupScript !== undefined) {
-        update.setup_script = options.setupScript || null;
-      }
-      if (options.cleanupScript !== undefined) {
-        update.cleanup_script = options.cleanupScript || null;
-      }
-      if (options.archiveScript !== undefined) {
-        update.archive_script = options.archiveScript || null;
-      }
-      if (options.copyFiles !== undefined) {
-        update.copy_files = options.copyFiles || null;
-      }
-      if (options.parallelSetup !== undefined) {
-        update.parallel_setup_script = options.parallelSetup;
-      }
-      if (options.devServerScript !== undefined) {
-        update.dev_server_script = options.devServerScript || null;
-      }
-      if (options.defaultTargetBranch !== undefined) {
-        update.default_target_branch = options.defaultTargetBranch || null;
-      }
-      if (options.defaultWorkingDir !== undefined) {
-        update.default_working_dir = options.defaultWorkingDir || null;
-      }
-
-      if (Object.keys(update).length === 0) {
-        console.log("No updates specified.");
-        return;
-      }
-
-      const client = await ApiClient.create();
-      const repoId = await getRepositoryId(id, client);
-      const repo = await client.updateRepo(repoId, update);
-      console.log(`Repository ${repo.id} updated.`);
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    if (options.displayName !== undefined) {
+      update.display_name = options.displayName || null;
     }
+    if (options.setupScript !== undefined) {
+      update.setup_script = options.setupScript || null;
+    }
+    if (options.cleanupScript !== undefined) {
+      update.cleanup_script = options.cleanupScript || null;
+    }
+    if (options.archiveScript !== undefined) {
+      update.archive_script = options.archiveScript || null;
+    }
+    if (options.copyFiles !== undefined) {
+      update.copy_files = options.copyFiles || null;
+    }
+    if (options.parallelSetup !== undefined) {
+      update.parallel_setup_script = options.parallelSetup;
+    }
+    if (options.devServerScript !== undefined) {
+      update.dev_server_script = options.devServerScript || null;
+    }
+    if (options.defaultTargetBranch !== undefined) {
+      update.default_target_branch = options.defaultTargetBranch || null;
+    }
+    if (options.defaultWorkingDir !== undefined) {
+      update.default_working_dir = options.defaultWorkingDir || null;
+    }
+
+    if (Object.keys(update).length === 0) {
+      console.log("No updates specified.");
+      return;
+    }
+
+    const client = await ApiClient.create();
+    const repoId = await getRepositoryId(id, client);
+    const repo = await client.updateRepo(repoId, update);
+    console.log(`Repository ${repo.id} updated.`);
   });
 
 // List branches for a repository
@@ -274,41 +246,35 @@ repositoryCommand
   .option("--local", "Show only local branches")
   .option("--json", "Output as JSON")
   .action(async (options, id?: string) => {
-    try {
-      const client = await ApiClient.create();
-      const repoId = await getRepositoryId(id, client);
-      let branches = await client.getRepoBranches(repoId);
+    const client = await ApiClient.create();
+    const repoId = await getRepositoryId(id, client);
+    let branches = await client.getRepoBranches(repoId);
 
-      // Filter by remote/local if specified
-      if (options.remote) {
-        branches = branches.filter((b) => b.is_remote);
-      } else if (options.local) {
-        branches = branches.filter((b) => !b.is_remote);
-      }
-
-      if (options.json) {
-        console.log(JSON.stringify(branches, null, 2));
-        return;
-      }
-
-      if (branches.length === 0) {
-        console.log("No branches found.");
-        return;
-      }
-
-      const table = new Table()
-        .header(["Name", "Current", "Remote"])
-        .body(
-          branches.map((b) => [
-            b.name,
-            b.is_current ? "*" : "",
-            b.is_remote ? "Yes" : "No",
-          ]),
-        );
-
-      table.render();
-    } catch (error) {
-      handleCliError(error);
-      throw error;
+    if (options.remote) {
+      branches = branches.filter((b) => b.is_remote);
+    } else if (options.local) {
+      branches = branches.filter((b) => !b.is_remote);
     }
+
+    if (options.json) {
+      console.log(JSON.stringify(branches, null, 2));
+      return;
+    }
+
+    if (branches.length === 0) {
+      console.log("No branches found.");
+      return;
+    }
+
+    const table = new Table()
+      .header(["Name", "Current", "Remote"])
+      .body(
+        branches.map((b) => [
+          b.name,
+          b.is_current ? "*" : "",
+          b.is_remote ? "Yes" : "No",
+        ]),
+      );
+
+    table.render();
   });
